ログイン機能とログイン済みユーザー専用の機能を実装

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -5,6 +5,9 @@ namespace App\Http\Controllers\Tweet;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\Tweet;   // Eloquentモデルを取得
+// テキストp139の内容
+use App\Services\TweetService;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;;
 
 class DeleteController extends Controller
 {
@@ -14,13 +17,17 @@ class DeleteController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
         $tweetId = (int) $request->route('tweetId');
+        // テキストp138の内容
+        if (!$tweetService->checkOwnTweet($request->user()->id, $tweetId)) {
+            throw new AccessDeniedHttpException();
+        }
         $tweet = Tweet::where('id', $tweetId)->firstorFail();
         $tweet->delete();
         return redirect()
-        ->route('tweet.index')
-        ->with('feedback.success', "つぶやきを削除しました。");
+            ->route('tweet.index')
+            ->with('feedback.success', "つぶやきを削除しました。");
     }
 }
